frontend: Rename ConfirmationContent.react.js to ConfirmationContent.tsx
