+++ frontend/src/js/__tests__/Common/ConfirmationContent.spec.js
@@ -1,6 +1,6 @@
 import {fireEvent, render} from '@testing-library/react';
 import React from 'react';
-import ConfirmationContent from '../../components/Common/ConfirmationContent.react';
+import ConfirmationContent from '../../components/Common/ConfirmationContent';
 import ApplicationsStore from '../../stores/ApplicationsStore';
 
 function mockResolver() {
