+++ frontend/src/js/components/Common/ConfirmationContent.tsx
